requirements.txt
asteval==0.9.32
blinker==1.7.0
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==2.8
Flask==2.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.14
pyjnius==1.6.1
PyJWT==2.8.0
requests==2.31.0
setuptools==51.3.3
urllib3==2.2.1
Werkzeug==2.2.2
